requirements.txt
awscli
braceexpand
easydict
ftfy
hydra-core
lightning-bolts==0.5.0
omegaconf
pickle5
pytorch-lightning==1.7.6
pycocotools
scikit-learn
torch>=1.10
torchvision>=0.11.0
tqdm
transformers
mypy-extensions
evalai
click==8.1.3
